feature(JobDefinitions): add job definitions section to process management
