+++ frontend/public/translations_en.json
@@ -204,6 +204,7 @@
 		},
 		"noResults": "No results",
 		"noIncidents": "No incidents",
+		"noJobDefinitions": "No job definitions",
 		"stacktrace": "Stacktrace",	
 		"copyValueToClipboard": "Copy value to clipboard",
 		"confirmRetryJob": "Are you sure to increment the number of retries?",
