+++ frontend/public/translations_es.json
@@ -162,6 +162,7 @@
 			"actions": "Acciones",
 			"active": "Activo",
 			"suspended": "Suspendido",
+			"activateJob": "Activar definición de trabajo",
 			"suspendJob": "Suspender definición de trabajo",
 			"activateMsg": "Esta definición de trabajo será activada, por lo que los nuevos trabajos basados en esta definición estarán activos inicialmente.",
 			"suspendMsg": "Esta definición de trabajo será suspendida, por lo que los nuevos trabajos basados en esta definición estarán suspendidos inicialmente.",
@@ -203,6 +204,7 @@
 		},
 		"noResults": "Sin resultados",
 		"noIncidents": "Sin incidencias",
+		"noJobDefinitions": "Sin definiciones de trabajos",
 		"stacktrace": "Traza de pila",	
 		"copyValueToClipboard": "Copiar valor al portapapeles",
 		"confirmRetryJob": "¿Desea incrementar el número de reintentos?",
