+++ frontend/public/translations_ru.json
@@ -204,6 +204,7 @@
 		},
 		"noResults": "Ничего не найдено",
 		"noIncidents": "Нет инцидентов",
+		"noJobDefinitions": "Нет определений заданий",
 		"stacktrace": "Стектрайс",	
 		"copyValueToClipboard": "Скопировать значение в буфер обмена",
 		"confirmRetryJob": "Вы действительно хотите увеличить число повторных попыток выполнения активности при неудачи?",
